Better filtering of visible rows, but the CSS fucks up
Doesn't look like there's a way to select only visible rows with CSS
Could append a <tr> after the removed row, but
  (a) super hacky
  (b) wouldn't be retained when reloading the page
      could add that into the page source, but that's hyper hacky

Probably the right thing to do is keep a separate hidden list of all the
applicants, and then copy the rows back and forth as they get filtered
in and out.

Also, there's some conflation of different ideas which is causing this
to be more difficult: Some things like "by date" aren't filters, but are
rather sorters. If we switch the implementation to use this style where
clicking a filter just adjusts which rows are in the presented list vs
the source list, then clicking a sorter just applies that organization
to the already filtered list.