Add graphviz file for rendering the state machine flow
For discussion: https://github.com/JumpstartLab/asquared/issues/43

This will allow us to edit the diagram as we edit the state machine
flow.

Currently the generated image is in .gitignore, b/c if it changes a lot,
then our repo could get big as its full of binary image data.